improvement: `nyl new component` now includes `metadata.labels` and `metadata.annotations` defaults in the generated `values.yaml` (#69)
This PR adds defaults for the `metadata.labels` and
`metadata.annotations` that Helm charts used as Nyl components can
received when instantiated by Nyl. When the instantiating resource has a
metadata field not set, it will not be passed. In order to simplify
writing Helm components, having them as empty default values works
around the need use the `default` function in some cases.